Fix all test warnings and clean up repository
- Fix module system to intelligently detect Dict[str, Module] types
- Fix DeprecationWarnings by updating to use ConfigManager correctly
- Fix JAX array static warnings with proper type annotations
- Clean up repository: remove temporary files, old test versions
- Update .gitignore to exclude backup and temporary files
- Add emb.field alias to avoid leaking equinox abstractions
- All 831 tests now pass with zero warnings

+++ tests/unit/operators/test_update_params.py
@@ -17,8 +17,8 @@ class SimpleOperator(operators.Operator):
     config: dict
 
     def __init__(self, dim: int):
-        self.weights = jnp.ones(dim)
-        self.bias = jnp.zeros(())
+        self.weights = jnp.ones(dim, dtype=jnp.float32)
+        self.bias = jnp.zeros((), dtype=jnp.float32)
         self.name = "test_op"
         self.config = {"dim": dim}
